% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compare_images.R
\name{compare_images}
\alias{compare_images}
\title{Compare image matches in the viewer window}
\usage{
compare_images(result, remove_duplicates = TRUE, previous = NULL)
}
\arguments{
\item{result}{TKTK}

\item{remove_duplicates}{TKTK}

\item{previous}{TKTK}
}
\value{
TKTK
}
\description{
\code{compare_images} TKTK
}
\details{
TKTK
}
