% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/confirm_matches.R
\name{confirm_matches}
\alias{confirm_matches}
\title{Confirm matches using a more detailed colour signature}
\usage{
confirm_matches(
  data,
  check_threshold = 0.99,
  confirm_thresholds = c(0.95, 0.98),
  quiet = FALSE
)
}
\arguments{
\item{data}{TKTK}

\item{check_threshold}{TKTK}

\item{confirm_thresholds}{TKTK}

\item{quiet}{TKTK}
}
\value{
TKTK
}
\description{
\code{confirm_matches} TKTK
}
\details{
TKTK
}
