% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_signature.R
\name{create_signature}
\alias{create_signature}
\alias{create_signature.matchr_img}
\alias{create_signature.cimg}
\alias{create_signature.character}
\alias{create_signature.list}
\alias{create_signature.default}
\title{Generate a unique colour signature for an image}
\usage{
create_signature(image, bands = 20, rm_black_bars = TRUE, ...)

\method{create_signature}{matchr_img}(image, bands = 20, rm_black_bars = TRUE, ...)

\method{create_signature}{cimg}(image, bands = 20, rm_black_bars = TRUE, ...)

\method{create_signature}{character}(
  image,
  bands = 20,
  rm_black_bars = TRUE,
  batch_size = 100,
  quiet = FALSE,
  ...
)

\method{create_signature}{list}(
  image,
  bands = 20,
  rm_black_bars = TRUE,
  batch_size = 100,
  quiet = FALSE,
  ...
)

\method{create_signature}{default}(image, bands = 20, rm_black_bars = TRUE, ...)
}
\arguments{
\item{image}{Object (or list of objects) of class \code{matchr_img} or \code{cimg},
probably imported using \code{load_image} or \code{imager::load.image}, or
file path or URL (or vector/list of file paths or URLs) which can be
imported to \code{matchr_img} using \code{load_image}.}

\item{bands}{Integer scalar. The number of horizontal and vertical bands the
image should be split into for processing. Higher values will produce a more
distinctive colour signature, potentially decreasing the rate of matching
false positives, but at the cost of increased processing time and an
increased rate of matching false negatives.}

\item{rm_black_bars}{Logical scalar. Should horizontal black bars be
detected and removed from the image signature? Because these bands lead to an
image signature dominated by black values, leaving them in the signature can
lead to false positive matches.}

\item{...}{Additional arguments passed to methods.}

\item{batch_size}{An integer scalar. How many images should the function
load into memory before extracting image signatures and releasing the
associated memory? Higher values will lead to the function executing more
quickly, but can result in enormous memory requirements.}

\item{quiet}{A logical scalar. Should the function execute quietly, or should
it return status updates throughout the function (default)?}
}
\value{
If the input is a single \code{matchr_img} or \code{cimg} image, an object of
class \code{matchr_sig}, which contains a numeric vector of length \code{bands * 8}, a
file name (optionally), and an aspect ratio. If the input is a character
vector of file paths or a list of \code{matchr_img} images, a list of class
\code{matchr_sig_list}.
}
\description{
\code{create_signature} takes an image file and generates a numeric vector of
average greyscale and colour values, so that the image can be compared to
others.
}
\details{
An image is decomposed into horizontal and vertical bands (with the number of
bands controlled by the \code{bands} argument), and for each band an average
greyscale and colour value is calculated. The vector of these averages
becomes a distinctive signature that can identify a given image even if the
image is rescaled or compressed, and thus serves as a reliable indicator of
whether two images are the same.
}
\examples{
# Import image with load_image then create signature
img <- load_image("https://upgo.lab.mcgill.ca/img/UPGo_logo.png")
create_signature(img)

# Or create signature directly from path/URL
create_signature("https://upgo.lab.mcgill.ca/img/UPGo_logo.png")
}
