% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/download_images.R
\name{download_images}
\alias{download_images}
\title{Download images to a local folder}
\usage{
download_images(
  x = NULL,
  destination,
  path = photos,
  id = id,
  check_duplicates = TRUE,
  quiet = FALSE
)
}
\arguments{
\item{x}{A data frame with a column containing URLs to be downloaded and
IDs to be linked to the downloads for file naming and duplicate detection.
If x is supplied, the \code{path} and \code{id} arguments will be interpreted as field
names. If x is NULL, they will be interpreted as lists or character vectors.}

\item{destination}{A character scalar indicating a local folder where images
should be downloaded to.}

\item{path}{If x is supplied, the name of a list or character field in x with
URLs to be downloaded. If x is NULL, a list or character vector of URLs. In
either case, elements should begin with "http", "https", "ftp" or "ftps". If
the URL has no extension, it will be given the extension ".jpg".}

\item{id}{If x is supplied, the name of a character field in x with IDs to
match to the file downloads. If x is NULL, a character vector the same length
as \code{path}, indicating the IDs of the files to be downloaded. Downloaded files
will be named according to the IDs, as \code{id_n.jpg}, where \code{id} is an element
of the \code{id} argument, \code{n} is an integer sequence of the same length as the
\code{id} element, and \code{.jpg} is the same file extension as the URL. (So if there
are three URLs which correspond to the \code{id} value of "XXXX", the files
downloaded will be "XXXX-1.jpg", "XXXX-2.jpg" and "XXXX-3.jpg".) If \code{id} has
repeating values, the numbering will be cumulatively sequential to avoid
overwriting earlier files, with, e.g., the second instance of "XXXX"
beginning with "XXXX-4.jpg".}

\item{check_duplicates}{A logical scalar (default TRUE) indicating whether to
skip downloads for files whose IDs match files already present in the
download folder. This check is done per ID rather than per file, which means
that if any files with an ID of \code{x} are found in the download folder, no new
files with an ID of \code{x} will be downloaded.}

\item{quiet}{A logical scalar. Should the function execute quietly, or should
it return status updates throughout the function (default)?}
}
\value{
Invisibly, a data frame summarizing the results of the downloads.
}
\description{
\code{download_images} takes a data frame or a vector of URLs and downloads
the files to a local folder, possibly with the use of an additional ID vector
to rename the files and check for duplicates.
}
\examples{
\dontrun{
# Setup
df <- data.frame(id = 1:15, photos = test_urls)
dest <- tempdir()
download_images(df, dest)
}
}
