% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check.R
\name{enum_type}
\alias{enum_type}
\title{Enum Type}
\usage{
enum_type(x, ...)
}
\arguments{
\item{x}{Enum object}

\item{...}{objects passed to methods}
}
\value{
character with the name of the enumerated type or \code{NULL}
}
\description{
Return the enumerated type name of an object, if a name was provided.
}
\examples{
x <- Result$Ok("hello world!")
enum_type(x)  # "Result"
}
