% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{get_*}
\alias{get_*}
\alias{get_path}
\alias{get_array}
\alias{get_signature}
\alias{get_x_sig}
\alias{get_y_sig}
\alias{get_ar}
\alias{get_x_ar}
\alias{get_y_ar}
\title{Get internal fields from matchr vectors}
\usage{
get_path(x)

get_array(x)

get_signature(x)

get_x_sig(x)

get_y_sig(x)

get_ar(x)

get_x_ar(x)

get_y_ar(x)
}
\arguments{
\item{x}{For \code{get_path} a \code{matchr_image} or \code{matchr_signature}
vector; for \code{get_array} a \code{matchr_image} or \code{matchr_matrix}
vector; for \code{get_signature} and \code{get_ar} a \code{matchr_signature}
vector; for \code{get_x_sig}, \code{get_y_sig}, \code{get_x_ar} and
\code{get_y_ar} a \code{matchr_matrix} vector.}
}
\value{
A list, with the same length as the input vector \code{x}, containing the
field \code{*} of the input vector corresponding to the \verb{get_*} function which was
called.
}
\description{
matchr vectors are implemented as "record"-style classes
(using \code{vctrs::new_rcrd}), which store their data and metadata as a set
of lists of equal length, called "fields". Usually these fields do not need
to be accessed independently of the vectors they comprise. But, for the cases
when they do, matchr has a series of \verb{get_*} functions which allow
convenient access to fields, and which are all straightforward wrappers of
\code{vctrs::field}. In each case, \code{get_*(x)} is equivalent to
\code{vctrs::field(x, "*")}.
}
\examples{
\dontrun{
# Setup
img <- load_image(test_urls)
sig <- create_signature(img)
matches <- match_signatures(sig)

get_path(img)
get_path(sig)
get_array(img)
get_array(matches)
get_signature(sig)
get_x_sig(matches)
get_y_sig(matches)
get_ar(sig)
get_x_ar(matches)
get_y_ar(matches)
}
}
