% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{get_ar}
\alias{get_ar}
\alias{get_x_ar}
\alias{get_y_ar}
\title{Get aspect ratios from a matchr_signature or matchr_matrix vector}
\usage{
get_ar(x)
}
\arguments{
\item{x}{TKTK}
}
\value{
TKTK
}
\description{
TKTK
}
\details{
TKTK
}
