% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{get_sig}
\alias{get_sig}
\title{Get image signatures from a matchr_matrix vector}
\usage{
get_sig(x, x_or_y)
}
\arguments{
\item{x}{TKTK}

\item{x_or_y}{TKTK}
}
\value{
TKTK
}
\description{
TKTK
}
\details{
TKTK
}
