% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fun.R
\name{\%.\%}
\alias{\%.\%}
\title{Compose Functions}
\usage{
lhs \%.\% rhs
}
\arguments{
\item{lhs}{function to be called second}

\item{rhs}{function to be called first}
}
\value{
a composed function
}
\description{
Combine two functions into a single function so that the \code{rhs} is called on the arguments first,
then the \code{lhs}.
}
\examples{
sq_log <- round \%.\% sqrt \%.\% log

Match(
  10:20,
  i \%fn\% (sq_log(i) > 2) ->
    "big",
  . ->
    "small"
)
}
