% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/matchr_helpers.R
\name{handler_matchr}
\alias{handler_matchr}
\title{Helper function to set a progress reporting strategy}
\usage{
handler_matchr(message)
}
\arguments{
\item{message}{The message to be displayed}
}
\value{
A progressr handler.
}
\description{
Helper function to set a progress reporting strategy
}
