% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/identify_image.R
\name{identify_image}
\alias{identify_image}
\title{Generate a unique colour signature for an image}
\usage{
identify_image(image, bands = 25)
}
\arguments{
\item{image}{Object of class 'cimg', probably imported using
\code{load_image} or \code{imager::load.image}, or file path or URL of an
image.}

\item{bands}{Integer scalar. The number of horizontal bands the image should
be split into for processing. Higher values will produce a more distinctive
colour signature, potentially decreasing the rate of matching false
positives, but at the cost of increased processing time and an increased rate
of matching false negatives.}
}
\value{
A numeric vector of length \code{bands}.
}
\description{
\code{identify_image} takes an image file and generates a numeric vector of
average colour values, so that the image can be compared to others.
}
\details{
The function works as follows. An image is decomposed into horizontal bands
(with the number of bands controlled by the \code{bands} argument), and for each
band an average colour is calculated. The vector of these colour averages
becomes a distinctive signature that can identify a given image even if the
image is rescaled or compressed, and thus serves as a reliable indicator of
whether two images are the same.
}
