% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/result_option.R
\name{into_option}
\alias{into_option}
\title{Convert Object into Option}
\usage{
into_option(x, ...)
}
\arguments{
\item{x}{Object to be converted}

\item{...}{Objects passed to methods}
}
\value{
an Enum object of class \code{Option}
}
\description{
Create an \code{\link{Option}} out of an object. By default the object is wrapped in a \code{Some} variant.
\code{Ok} variants of \code{\link{Result}} are turned into \code{Some} Options, while \code{Err} variants are
turned into \code{None} Options.
}
\examples{
an_error <- Result$Err("hello world!")
into_option(an_error) # None

}
