% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/result_option.R
\name{into_result}
\alias{into_result}
\title{Convert Object into Result}
\usage{
into_result(x, ...)
}
\arguments{
\item{x}{Object to be converted}

\item{...}{Objects passed to methods}
}
\value{
an Enum object of class \code{Result}
}
\description{
Create a \code{\link{Result}} out of an object. By default the object is wrapped in an \code{Ok} variant.
\code{Some} variants of \code{\link{Option}} are turned into \code{Ok} Results, while \code{None} variants are
turned into \code{Err} Results
}
\examples{
nothing <- Option$None
into_result(nothing) # Err

}
