% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check.R
\name{is.enum}
\alias{is.enum}
\title{Is Object an Enum}
\usage{
is.enum(x)
}
\arguments{
\item{x}{object to be tested}
}
\value{
\code{TRUE} if \code{x} is an Enum, \code{FALSE} otherwise
}
\description{
Test whether object has class \code{\link{Enum}}.
}
\examples{
HelloEnum <- Enum(
  "HelloEnum",
  Hello,
  World
)

# TRUE
is.enum(HelloEnum$Hello)

# FALSE
is.enum(5)

}
