% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check.R
\name{is.none}
\alias{is.none}
\title{Check if Option is None}
\usage{
is.none(x)
}
\arguments{
\item{x}{object to be tested}
}
\value{
\code{TRUE} if \code{x} is enumerated type of variant \code{None}, \code{FALSE} otherwise
}
\description{
Test whether Option Enum is Some or None.
}
\examples{
x <- 1:5
get_n <- function(x, n) {
  if (n > length(x)) return(None)
  Some(x[n])
}
obj <- get_n(x, 6)
is.none(obj)  # TRUE
}
