% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check.R
\name{is.ok}
\alias{is.ok}
\title{Check if Result is Ok}
\usage{
is.ok(x)
}
\arguments{
\item{x}{object to be tested}
}
\value{
\code{TRUE} if \code{x} is enumerated type of variant \code{Ok}, \code{FALSE} otherwise
}
\description{
Test whether Result Enum is Ok or an Err.
}
\examples{
sqrt_big <- function(x) {
  if (x > 1000){ return(Ok(sqrt(x))) }
  Err("Not large enough!")
}
x <- sqrt_big(250)
is.ok(x)  # FALSE

}
