% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/matchr_image.R
\name{is_image}
\alias{is_image}
\title{Test if the object is a matchr_image}
\usage{
is_image(x)
}
\arguments{
\item{x}{An object to test}
}
\value{
A logical scalar, TRUE if \code{x} inherits from class "matchr_image" and
FALSE otherwise.
}
\description{
This function returns TRUE for \code{matchr_image} objects and FALSE for all other
objects.
}
\examples{
\dontrun{
# Setup
img <- load_image(test_urls)

# TRUE
is_image(img)

# FALSE
is_image("text")
}
}
