% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/matchr_matrix.R
\name{is_matrix}
\alias{is_matrix}
\title{Test if the object is a matchr_matrix}
\usage{
is_matrix(x)
}
\arguments{
\item{x}{An object to test}
}
\value{
A logical scalar, TRUE if \code{x} inherits from class "matchr_matrix" and
FALSE otherwise.
}
\description{
This function returns TRUE for \code{matchr_matrix} objects and FALSE for all
other objects.
}
