% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/matchr_signature.R
\name{is_signature}
\alias{is_signature}
\title{Test if the object is a matchr_signature}
\usage{
is_signature(x)
}
\arguments{
\item{x}{An object to test}
}
\value{
A logical scalar, TRUE if \code{x} inherits from class "matchr_signature"
and FALSE otherwise.
}
\description{
This function returns TRUE for \code{matchr_signature} objects and FALSE for all
other objects.
}
