% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/matchr_signature_2.R
\name{is_signature_2}
\alias{is_signature_2}
\title{Test if the object is a matchr_signature}
\usage{
is_signature_2(x)
}
\arguments{
\item{x}{An object to test}
}
\value{
A logical scalar, TRUE if \code{x} inherits from class "matchr_signature_2"
and FALSE otherwise.
}
\description{
This function returns TRUE for \code{matchr_signature_2} objects and FALSE for all
other objects.
}
\examples{
\dontrun{
# Setup
sigs <- create_signature(test_urls)

# TRUE
is_signature(sigs)

# FALSE
is_signature("text")
}
}
