% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load_image.R
\name{load_image}
\alias{load_image}
\title{Load image from file or URL}
\usage{
load_image(file, quiet = FALSE)
}
\arguments{
\item{file}{A vector of file paths or URLs to be passed to
\code{imager::load.image}. If \code{file} is a vector of URLs, the URLs must begin
with "http", "https", "ftp" or "ftps". If the URL has no extension, it will
be given the extension ".jpg".}

\item{quiet}{A logical scalar. Should the function execute quietly, or should
it return status updates throughout the function (default)?}
}
\value{
A list of \code{matchr_img} objects of the same length as the input
vector.
}
\description{
\code{load_image} is a vectorized version of \code{imager::load.image}. It
optionally supports parallel processing (via \code{future} and
\code{future.apply}) and progress reporting.
}
\details{
The function is a wrapper around \code{imager::load.image} with five
enhancements. It can take a vector of input paths instead a single path; it
supports parallel processing; it support progress reporting; it appends the
file path of the image as an attribute for subsequent processing or record
keeping; and it suppresses file download messages to allow more meaningful
progress reporting.
}
