% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/match.R
\name{match_cond}
\alias{match_cond}
\title{Check and Evaluate Match Condition}
\usage{
match_cond(cond, x, do, ...)
}
\arguments{
\item{cond}{match condition}

\item{x}{object being matched}

\item{do}{return expression associated with the condition. If \code{cond} is matched with \code{x}, then \code{do}
should be evaluated and returned in a list with \code{TRUE}: \code{list(TRUE, eval(do))}.}

\item{...}{arguments passed to evaluation}
}
\value{
\code{FALSE} if no match, or a list containing \code{TRUE} and the evaluated expression
}
\description{
Called by \code{\link{Match}} the check whether a condition matches. Used to create custom methods for matching.
}
\details{
See the \code{\link{Match}} details for explanations about provided methods.
}
