% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/matchr.R
\name{matching_test}
\alias{matching_test}
\title{Matching test between two datasets.}
\usage{
matching_test(data1, data2, id1 = data1$id, id2 = data2$id)
}
\arguments{
\item{data1}{data.frame. One of two datasets.}

\item{data2}{data.frame. The other dataset.}

\item{id1}{data1's id. This parameter should be integer.}

\item{id2}{data2's id. This parameter should be integer.}
}
\value{
Output on console about missmatch variable and participant.
}
\description{
Matching test between two datasets.
}
\note{
This package support only numeric or integer variable. A factor type variable will be supported later update.
}
\examples{
n <- 20
min <- 0
max <- 5
id <- 1:n
id2 <- n:1
set.seed(123)
var1 <- runif(n = n, min = min, max = max)
set.seed(213)
var2 <- runif(n = n, min = min, max = max)
set.seed(321)
var3 <- runif(n = n, min = min, max = max)
hoge1 <- data.frame(key = id, var1, var2, var3)
hoge2 <- data.frame(id, var1, var2, var3)
hoge2$var1[2] <- 6.00
hoge2$var2[7] <- -1.00
hoge2$var3[8] <- NA
data3 <- dplyr::arrange(hoge2, -id)#data3 reversed order data2 dataset

matching_test(data1 = hoge1, data2 = hoge2, id1 = hoge1$key, id2 = hoge2$id)
matching_test(data1 = hoge1, data2 = data3, id1 = hoge1$key, id2 = data3$id)

}
