% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/matchr-package.R
\docType{package}
\name{matchr-package}
\alias{matchr}
\alias{matchr-package}
\title{matchr: Fast and Reliable Matching of Images}
\description{
Fast and reliable matching of images. It creates unique image signatures using perceptual hashing. This allows large sets of images to be compared to each other to identify image matches. The matching process is insensitive to file format, resolution, or compression.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/UPGo-McGill/matchr}
  \item Report bugs at \url{https://github.com/UPGo-McGill/matchr/issues}
}

}
\author{
\strong{Maintainer}: David Wachsmuth \email{david.wachsmuth@mcgill.ca} (\href{http://orcid.org/0000-0001-5689-9527}{ORCID})

}
\keyword{internal}
