% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/matchr_options.R
\name{matchr_options}
\alias{matchr_options}
\alias{matchr.options}
\alias{matchr.blas}
\alias{matchr.force_parallel}
\title{Options used in matchr}
\description{
Below are the R options and environment variables that are used by the matchr
package.
}
\section{matchr options}{


\describe{
\item{\option{matchr.blas}:}{A logical scalar. TKTK. (Default: \code{TRUE})}
}

\describe{
\item{\option{matchr.force_parallel}:}{A logical scalar. TKTK (Default:
\code{FALSE})}
}
}

