% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/multicolumn_join.R
\name{multicolumn_join}
\alias{multicolumn_join}
\title{Perform a Left Join Based on Multi-Column Matching}
\usage{
multicolumn_join(x, xcols, y, ycols)
}
\arguments{
\item{x}{The "left" data frame (the data frame to which \code{y} will be joined).}

\item{xcols}{A character vector of column names in \code{x} to use for matching.}

\item{y}{The "right" data frame (the data frame from which columns will be
joined).}

\item{ycols}{A character vector of column names in \code{y} to use for matching.}
}
\value{
A new data frame resulting from the left join of \code{x} and \code{y}.
Rows in \code{x} that do not have a match in \code{y} will have \code{NA} values for
columns originating from \code{y}.
}
\description{
This function performs a left join operation on two data frames (\code{x} and \code{y})
based on multi-column matching. It first uses \code{multicolumn_match()} to
identify corresponding rows and then joins the data frames using a temporary
index column.
}
\examples{
\dontrun{
# Assuming 'df1' and 'df2' are your data frames
# and 'df1_gene_cols', 'df2_gene_cols' are the relevant column names
# joined_df <- multicolumn_join(df1, df1_gene_cols, df2, df2_gene_cols)
}
}
