% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/matchr_image.R
\name{new_image}
\alias{new_image}
\title{Create a new matchr_image object}
\usage{
new_image(x = list(), file = character())
}
\arguments{
\item{x}{A list of pixel arrays.}

\item{file}{A character string, corresponding to the path or URL of the files
from which the arrays have been generated.}
}
\value{
An object of class \code{matchr_image}.
}
\description{
Create a new matchr_image object
}
