% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/class_matchr_change_table.R
\name{new_matchr_change_table}
\alias{new_matchr_change_table}
\title{Create a new matchr_change_table object}
\usage{
new_matchr_change_table(x)
}
\arguments{
\item{x}{A data frame.}
}
\value{
An data frame of class \code{matchr_change_table}.
}
\description{
Create a new matchr_change_table object
}
