% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/class_matchr_img.R
\name{new_matchr_img}
\alias{new_matchr_img}
\title{Create a new matchr_img object}
\usage{
new_matchr_img(x, file)
}
\arguments{
\item{x}{An array.}

\item{file}{A character string, corresponding to the path or URL of the file
from which the image has been generated.}
}
\value{
An object of class \code{matchr_img}.
}
\description{
Create a new matchr_img object
}
