% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/class_matchr_matrix.R
\name{new_matchr_matrix}
\alias{new_matchr_matrix}
\title{Create a new matchr_matrix object}
\usage{
new_matchr_matrix(x, x_aspect_ratio, y_aspect_ratio)
}
\arguments{
\item{x}{A matrix of correlations.}

\item{x_aspect_ratio}{A numeric scalar.}

\item{y_aspect_ratio}{A numeric scalar.}
}
\value{
An object of class \code{matchr_matrix}.
}
\description{
Create a new matchr_matrix object
}
