% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/class_matchr_matrix_list.R
\name{new_matchr_matrix_list}
\alias{new_matchr_matrix_list}
\title{Create a new matchr_matrix_list object}
\usage{
new_matchr_matrix_list(x, x_files, y_files)
}
\arguments{
\item{x}{A list of matchr_matrix objects.}

\item{x_files}{A numeric scalar: the number of files in the matrix rows.}

\item{y_files}{A numeric scalar: the number of files in the matrix columns}
}
\value{
An object of class \code{matchr_matrix_list}.
}
\description{
Create a new matchr_matrix_list object
}
