% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/class_constructors.R
\name{new_matchr_sig_list}
\alias{new_matchr_sig_list}
\title{Create a new matchr_sig_list object}
\usage{
new_matchr_sig_list(x)
}
\arguments{
\item{x}{A list of matchr_sig objects.}
}
\value{
An object of class \code{matchr_sig_list}.
}
\description{
Create a new matchr_sig_list object
}
