% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/matchr_matrix_2.R
\name{new_matrix_2}
\alias{new_matrix_2}
\title{Create a new matchr_matrix object}
\usage{
new_matrix_2(
  array = list(),
  x_ar = list(),
  y_ar = list(),
  x_sig = list(),
  y_sig = list(),
  x_total = integer(length = 1L),
  y_total = integer(length = 1L),
  x_na = character(),
  y_na = character()
)
}
\arguments{
\item{array}{A list of Hamming distance matrices.}

\item{x_ar}{A list of numeric vectors: the highest and lowest aspect ratio in
the x vector.}

\item{y_ar}{A list of numeric vectors: the highest and lowest aspect ratio in
the y vector.}

\item{x_sig}{A list of \code{matchr_signature} vectors: the signatures of the
files in the x vector.}

\item{y_sig}{A list of \code{matchr_signature} vectors: the signatures of the
files in the y vector.}

\item{x_total}{An integer scalar: the total number of x signatures analyzed.}

\item{y_total}{An integer scalar: the total number of y signatures analyzed.}

\item{x_na}{A character vector: the paths of x signatures which are NA.}

\item{y_na}{A character vector: the paths of y signatures which are NA.}
}
\value{
An object of class \code{matchr_matrix_2}.
}
\description{
Create a new matchr_matrix object
}
