% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/matchr_signature.R
\name{new_signature}
\alias{new_signature}
\title{Create a new matchr_signature object}
\usage{
new_signature(x = list(), path = character(), aspect_ratio = numeric())
}
\arguments{
\item{x}{A list of numeric vectors.}

\item{path}{A character string, corresponding to the path or URL of the files
from which the signatures have been generated.}

\item{aspect_ratio}{A numeric vector, giving the aspect ratio of the images.}
}
\value{
An object of class \code{matchr_signature}.
}
\description{
Create a new matchr_signature object
}
