% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/matchr_signature_2.R
\name{new_signature_2}
\alias{new_signature_2}
\title{Create a new matchr_signature object}
\usage{
new_signature_2(x = list(), path = character(), ar = numeric())
}
\arguments{
\item{x}{A list of numeric vectors.}

\item{path}{A character string, corresponding to the path or URL of the files
from which the signatures have been generated.}

\item{ar}{A numeric vector, giving the aspect ratio of the images.}
}
\value{
An object of class \code{matchr_signature_2}.
}
\description{
Create a new matchr_signature object
}
