% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/result_option.R
\docType{data}
\name{None}
\alias{None}
\title{None}
\format{
an empty list of classes \code{"Option"} and \code{"Enum"}
}
\usage{
None
}
\description{
An Enum variant of \code{Option} used to denote that a function returned
no value.
}
\keyword{datasets}
