% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/normalize_names.R
\name{normalize_names}
\alias{normalize_names}
\title{Clean and Normalize Gene Names}
\usage{
normalize_names(
  names,
  to_sub = "[-.]",
  sub_with = "_",
  rm_suffix = FALSE,
  suffixes = c("_frag$", "-[0-9]$"),
  case_fun = "toupper"
)
}
\arguments{
\item{names}{A character vector of names to be normalized.}

\item{to_sub}{A regular expression specifying characters to be replaced
(default: \code{"[-.]"}, matching hyphens and periods).}

\item{sub_with}{The character to substitute for \code{to_sub} (default: \code{"_"}).}

\item{rm_suffix}{A logical value indicating whether to remove specified
suffixes (default: \code{FALSE}).}

\item{suffixes}{A character vector of regular expressions defining suffixes
to be removed if \code{rm_suffix} is \code{TRUE} (default: \code{c("_frag$","-[0-9]$")}).}

\item{case_fun}{A character string specifying a case conversion function
(e.g., \code{"toupper"}, \code{"tolower"}) to apply to the names. Use \code{NA} to skip
case conversion (default: \code{"toupper"}).}
}
\value{
A character vector of normalized names.
}
\description{
This function cleans and normalizes a character vector of names by performing
substitutions, optionally removing specified suffixes, and converting the
case of the names.
}
\examples{
normalize_names(c("Pe-Pe_frag", "nene.ne-2-0"))
normalize_names(c("GeneA", "GeneB_frag"), rm_suffix=TRUE)
normalize_names(c("GeneC", "GeneD"), case_fun="tolower")
}
