% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/one_factor_search.R
\name{one_factor_search}
\alias{one_factor_search}
\title{one_factor_search}
\usage{
one_factor_search(raw_tests, raw_controls)
}
\arguments{
\item{raw_tests}{The data.frame containing all of the tests}

\item{raw_controls}{The data.frame containing all of the controls}
}
\description{
Takes test group and control group (each as 3 column data.frames) with each
observation being defined by one categorical value and one value from a continuous 
variable. Performs 1-1 matching on both factor and continuous values.
Begins with first test observation, choosing closest match, and removing
it from consideration. Moves down the list, continuing the selection process.
Then performs the process in reverse, starting at the last test observation and
working up. The forward and backward selections are evaluated based upon which
has the lower sum of absolute deviations.
}
\details{
The tests are first parameter, and the controls are second. Make sure that the
order of each data.frame is FACTOR, OBSERVATION ID, and CONTINUOUS VALUE.
}
\examples{
raw_tests <- read.csv("raw_tests.csv")
raw_controls <- read.csv("raw_controls.csv")
two_factor_search(raw_tests, raw_controls) 
}
\author{
Nicholas Sun <nicholas.sun@rutgers.edu>
}

