% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/result_option.R
\docType{data}
\name{Option}
\alias{Option}
\title{Option}
\format{
list with 1 Enum generators and 1 Enum variant
\describe{
  \item{Some(x)}{Wrap \code{x} in the 'Some' variant.}
  \item{None}{Variant denoting that nothing was returned.}
}
}
\usage{
Option
}
\description{
An Enum that mimics Rust's "Option" type. This is used to denote whether
a function returned an object or not, rather than returning \code{NULL}.
}
\keyword{datasets}
