% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{plot.matchr_image}
\alias{plot.matchr_image}
\title{Plot a matchr_image vector}
\usage{
\method{plot}{matchr_image}(x, path = TRUE, max_plot = 12, n_rows = "auto", ...)
}
\arguments{
\item{x}{Vector of class \code{matchr_image}}

\item{path}{Logical scalar. Should the image's path be displayed above the
plot (default TRUE)?}

\item{max_plot}{Positive integer. The maximum number of images to plot at
once.}

\item{n_rows}{Either "auto" or a positive integer. The number of rows with
which to plot images. If "auto", six or fewer images will be arranged in an
adaptive layout which maximizes legibility, while seven or more images will
always be arranged in a three-column grid.}

\item{...}{Not used.}
}
\description{
Plot a matchr_image vector
}
