% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/matchr_signature.R
\name{plot.matchr_signature}
\alias{plot.matchr_signature}
\title{Plot a matchr_signature vector}
\usage{
\method{plot}{matchr_signature}(x, max_plot = 4, with_image = TRUE, ...)
}
\arguments{
\item{x}{Vector of class \code{matchr_signature}}

\item{max_plot}{Positive integer. The maximum number of signatures to plot at
once.}

\item{with_image}{Logical scalar. Should the underlying image be plotted
alongside its signature (default TRUE)? If TRUE, \code{\link{load_image}}
will be called on the paths in \code{x}.}

\item{...}{Not used.}
}
\description{
Plot a matchr_signature vector
}
