% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/remove_backups.R
\name{remove_backups}
\alias{remove_backups}
\title{Remove matchr function backups}
\usage{
remove_backups()
}
\value{
A character vector of the backups removed, invisibly.
}
\description{
\code{remove_backups} manually removes the incremental backups which some
matchr functions keep in case of errors.
}
\details{
The \code{\link{create_signature}} function can create incremental backups
of its progress in case it fails to complete. These backups are automatically
removed once the function completes, but \code{remove_backups} can be used to
remove them manually when necessary.
}
