% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/remove_black_bars.R
\name{remove_black_bars}
\alias{remove_black_bars}
\title{Remove horizontal black bars from an image}
\usage{
remove_black_bars(image, bands)
}
\arguments{
\item{image}{Object of class 'cimg', probably imported using
\code{load_image} or \code{imager::load.image}.}

\item{bands}{Integer scalar. The number of horizontal bands the image should
be split into for processing.}
}
\value{
An object of class \code{cimg}.
}
\description{
\code{remove_black_bars} takes an image file and removes black bars at the
top and bottom of the image, if any are identified.
}
\details{
TKTK
}
