% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/result_option.R
\docType{data}
\name{Result}
\alias{Result}
\title{Result}
\format{
list with 2 Enum generators
\describe{
  \item{Ok(x)}{Wrap \code{x} in the 'Ok' variant.}
  \item{Err(e)}{Wrap \code{x} in the 'Err' variant.}
}
}
\usage{
Result
}
\description{
An Enum that mimics Rust's "Result" type. This is used to denote whether
a function contained an error without stopping execution and allowing the
error result to be unwrapped.
}
\keyword{datasets}
