% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/matchr_helpers.R
\name{set_par}
\alias{set_par}
\title{Helper function to decide on future plan}
\usage{
set_par(fun, ...)
}
\arguments{
\item{fun}{A character string indicating the parent function}

\item{...}{Additional named arguments}
}
\value{
A logical scalar.
}
\description{
Helper function to decide on future plan
}
