% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{sig_length}
\alias{sig_length}
\title{Get the signature length from a matchr_signature vector}
\usage{
sig_length(x)
}
\arguments{
\item{x}{A matchr_signature vector.}
}
\value{
An integer scalar.
}
\description{
By default the raw image signatures in a matchr_signature vector have a
length of 160 (20 horizontal and vertical bands of greyscale, red, green, and
blue). But since the length can take any arbitrary value, \code{sig_length}
offers a convenient way to access this value on a per-vector basis.
}
