% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/subset_dim.R
\name{subset_dim}
\alias{subset_dim}
\title{subset_dim}
\usage{
subset_dim(raw_tests, raw_controls, num_cat_factors = 1)
}
\arguments{
\item{raw_tests}{The imported test data}

\item{raw_controls}{The imported control data}

\item{num_cat_factors}{The number of categorical factors we are subsetting on}
}
\description{
Gives the dimensions of each subset of the data; allows the analyst
to know the distribution of the tests after being divided along
one or two cateogorical factors, as well as the possible control pool
for each subset. A low ratio between tests and controls might indicate
that R was forced to make selections for certain controls that might not
be a good fit.
}
\details{
The number of categorical factors defaults to 1. 
categorical factors are in the leftmost columns of each data set.
}
\examples{
raw_tests <- read.csv("raw_tests_2_categories.csv")
raw_controls <- read.csv("raw_controls_2_categories.csv")
subset_dim(raw_tests, raw_controls, 2)
}
\author{
Nicholas Sun <nicholas.sun@rutgers.edu>
}

