% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/result_option.R
\name{Try}
\alias{Try}
\title{Execute Expression as Result}
\usage{
Try(expr)
}
\arguments{
\item{expr}{expression to evaluate}
}
\value{
Result Enum of variant \code{Ok} or \code{Err}
}
\description{
Evaluates given expression returning an \code{Err} \link{Result} if there is an error, otherwise an \code{Ok} Result.
}
\examples{
# This returns an Err
Try(sqrt + 1)

# This returns an Ok
Try(sqrt(5) + 1)
}
