% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/two_factor_search.R
\name{two_factor_search}
\alias{two_factor_search}
\title{two_factor_search}
\usage{
two_factor_search(raw_tests, raw_controls)
}
\arguments{
\item{raw_tests}{The data.frame containing the tests}

\item{raw_controls}{The data.frame containing the controls}
}
\description{
Takes test group and control group (as 4 column data.frame) with 
exactly two categorical variables and one continuous variable for each
obsevation. Performs 1-1 matching on both factors and a continuous value.
Begins with first test, removing selected controls as it travels down.
Then reperforms search going from the last test up to the first test.
Forwards and backwards selections evaluated based on which has lower sum
of absolute deviations from continuous values.
}
\details{
The tests are the first parameter, controls are second.
Make sure that order in each data.frame input is FIRST FACTOR,
SECOND FACTOR, OBSERVATIOIN ID, and CONTINUOUS VALUE.
}
\examples{
raw_tests <- read.csv("raw_tests.csv")
raw_controls <- read.csv("raw_controls.csv")
two_factor_search(raw_tests, raw_controls)
}
\author{
Nicholas Sun <nicholas.sun@rutgers.edu>
}

