% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/validate.R
\name{validate}
\alias{validate}
\title{validate}
\usage{
validate(output_df)
}
\arguments{
\item{output_df}{The data frame that contains all tests and selected controls}
}
\description{
Provides some diagnostics on the goodness of control fit to the test groups:
average difference, proportion of difference to test group average, and if 
there are any duplicats present in the controls.
}
\details{
If there are duplicates in the controls, the functions output will be a vector
of row numbers which contain the duplicated controls.
}
\examples{
output <- one_factor_search(raw_tests, raw_controls)
validate(output)
}
\author{
Nicholas Sun <nicholas.sun@rutgers.edu>
}

