% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/matchr_helpers.R
\name{chunk}
\alias{chunk}
\title{Helper function to divide vector into equal-sized chunks}
\usage{
chunk(x, n)
}
\arguments{
\item{x}{The vector to be split}

\item{n}{The number of chunks to split into}
}
\value{
A list with n elements.
}
\description{
Helper function to divide vector into equal-sized chunks
}
