% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compare_images.R
\name{compare_images}
\alias{compare_images}
\title{Compare image pairs in a Shiny app}
\usage{
compare_images(
  x, 
  y, 
  match = NULL, 
  batch_size = 100L,
  match_defaults = c(Match = TRUE, `Likely match` = TRUE, 
    `Possible match` = FALSE, `No match` = FALSE), 
  info = NULL, 
  ...
)
}
\arguments{
\item{x, y}{A \code{matchr_image}, \code{matchr_signature} or character vector which
contains references to a set of images. If a \code{matchr_image} vector, images
will be drawn directly from the in-memory pixel information in the vector.
If a \code{matchr_signature} vector, the images will be retrieved from the file
paths or URLs contained in each \code{matchr_signature} element. If a character
vector, the vector elements will be assumed to be file paths or URLs, from
which the images will be retrieved.}

\item{match}{An optional character vector. Image pairs will be grouped by
these values. (If there are more than 10 unique values, the function will
issue a warning, since this probably implies that the wrong vector was
passed to the argument.)}

\item{batch_size}{An integer scalar. The number of images to display at a
time in the Shiny app (default 100).}

\item{match_defaults}{A named logical vector. Rows whose \code{match} argument
value corresponds to a name in this vector will have a starting status of
"match" in the comparison app if the named element's value is TRUE and "no
match" if it is FALSE. If this argument is NULL, all image pairs will have a
starting status of "match".}

\item{info}{An optional data frame with information to be shown at the top of
the comparison app.}

\item{...}{Additional named vectors which will be displayed per-image pair
when the "Show details" button is toggled in the comparison app. These
vectors must be the same length as the \code{x} and \code{y} inputs, and their contents
will be truncated to fit in the comparison app, so they should be relatively
terse. Names are only to identify \code{...} arguments; they will not be
displayed alongside the vector contents. If multiple arguments to \code{...} are
supplied, they will be concatenated with "; ".}
}
\value{
A data frame with the following fields: \code{x} and \code{y} (or whatever
these fields were originally named) from the original \code{df} data frame;
\code{new_match_status}, which is a character vector with possible entries "match"
and "no match"; and a logical vector \code{new_highlight} which is TRUE for any
matches which were highlighted using the in-app interface. The output will
only have non-missing \code{new_match_status} values for each image pairing that
confirmed, which is determined by how many pages into the Shiny app the user
proceeded, and thus how many pairings were viewed. Rows with pairings which
were not viewed will have an NA value in \code{new_match_status} and
\code{new_highlight}. If all pages are viewed, then the output will have no NA
values.
}
\description{
\code{compare_images} takes two sets of images and displays them in pairs
for comparison.
}
\details{
The interface presents pairs of images alongside a best guess as to the match
status ("Match" or "No match"). For matches which are correctly identified,
no further action is necessary, while incorrect identifications can be
corrected by clicking "Match" or "No match" next to the image pair. Images
are presented in batches, and at any point the user can click the "Save and
exit" button to close the comparison app and retrieve the results up through
the last batch which was viewed. This means that even extremely large sets of
potential matches can be manually verified over the course of several
sessions. Through the "Enable highlighting" button, specific matches can be
highlighted for further follow-up after image comparison is finished.

The \code{match} field can contain any character strings, and image pairs will be
grouped by these values. (If there are more than 10 unique values for
\code{match}, the function will issue a warning, since this probably implies that
the \code{match} field was misidentified.) By default, all image pairs will start
with a value of "match" in the comparison app. But if the values of the
\code{match} field correspond to a name in the function's "match" argument, then
the value will be "match" if the corresponding "match" argument value is
TRUE, and "no match" if the corresponding "match" argument value is FALSE.
For example, by default the "match" argument is:
\verb{c("Match" = TRUE, "Likely match" = TRUE, "Possible match" = FALSE}.
So any image pairs with a \code{match} field value of "Match" or "Likely match"
will start with a value of "match" in the comparison app, while any image
pairs with a \code{match} field value of "Possible match" will start with a value
of "no match" in the comparison app.
}
\examples{
\dontrun{
# Assign the output of compare_images to retrieve results
change_table <- compare_images(test_urls, test_urls)
}
}
