% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compare_images.R
\name{compare_images}
\alias{compare_images}
\title{Compare image matches in a Shiny app}
\usage{
compare_images(
  result,
  remove_duplicates = TRUE,
  batch_size = 100L,
  corr_thresh = 0.9995,
  previous = TRUE,
  quiet = FALSE
)
}
\arguments{
\item{result}{A data frame produced by \code{\link{identify_matches}} (with
confirm = TRUE or with \code{\link{confirm_matches}} subsequently run on the
results). Any additional fields added to the data frame will be preserved in
the output.}

\item{remove_duplicates}{A logical scalar. Should x-y pairs which are
identical to other x-y pairs be reduced to a single x-y pair? This step can
be computationally expensive for large datasets, but can dramatically reduce
the number of matches to be verified.}

\item{batch_size}{An integer scalar. The number of images to display at a
time in the Shiny app (default 100).}

\item{corr_thresh}{A numeric scalar between 0 and 1. Matches with signature
correlations above this value (default 0.9995) will be pre-emptively marked
as "match". If \code{remove_duplicates} is TRUE, this same value will be used to
identify duplicated images. (I.e. if the correlation between two x or two y
images is >= \code{corr_thresh}, the images will be considered duplicates.) If
\code{corr_thresh} is set to 1, all image matches will be presented for manual
verification unless they are bit-for-bit identical.}

\item{previous}{A logical scalar. Should the results of previous runs of
\code{compare_images} be incorporated into the new results (default TRUE), or
should previously compared matches be compared again? If this argument is
TRUE, then any rows in \code{result} with a \code{confirmed} value of TRUE will be
removed from the data frame before processing (and so will not be present
in the comparison interface) and then re-added unchanged to the output.}

\item{quiet}{A logical scalar. Should the function execute quietly, or should
it return status updates throughout the function (default)?}
}
\value{
A data frame with the same fields as the input \code{result}, except 1)
the field \code{match} will be replaced with \code{new_match_status}, which is a
character vector with possible entries "match" and "no match"; and 2) a
logical field \code{highlight} will be added if any matches were highlighted using
the in-app interface. The output will have one row for each image pairing
that was confirmed, which is determined by how many pages into the Shiny app
the user proceeded, and thus how many pairings were viewed. If all pages are
viewed, then the output will have the same number of rows as the input.
}
\description{
\code{compare_images} takes the image matches produced by
\code{\link{identify_matches}} and \code{\link{confirm_matches}} and displays
them in an interactive Shiny app for visual inspection and confirmation.
Image matches with extremely high correlation can be optionally excluded, and
pairwise duplicates can be detected and excluded as well.
}
\details{
The interface presents pairs of images alongside a best guess as to the match
status ("Match" or "No match"). For matches which are correctly identified,
no further action is necessary, while incorrect identifications can be
corrected by clicking "Match" or "No match" next to the image pair. Images
are presented in batches, and at any point the user can click the "Save and
exit" button to close the comparison app and retrieve the results up through
the last batch which was viewed. This means that even extremely large sets of
potential matches can be manually verified over the course of several
sessions.

Through the "Enable highlighting" button, specific matches can be highlighted
for further follow-up after image comparison is finished.
}
\examples{
\dontrun{
# Setup
sigs <- create_signature(test_urls)
matches <- match_signatures(sigs)
confirm <- identify_matches(matches)

# Assign the output of compare_images to retrieve results
change_table <- compare_images(confirm)
}
}
