% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/confirm_matches.R
\name{confirm_matches}
\alias{confirm_matches}
\title{Confirm matches using a more detailed colour signature}
\usage{
confirm_matches(
  data,
  check_threshold = 0.99,
  confirm_thresholds = c(0.95, 0.98),
  quiet = FALSE,
  ...
)
}
\arguments{
\item{data}{A data frame produced from \code{\link{identify_matches}}.}

\item{check_threshold}{A numeric scalar. Matches with correlations higher
than this value (default 0.99) will be treated as matching and not
re-checked.}

\item{confirm_thresholds}{A numeric vector of length 2. New correlation
coefficients below the first value (default 0.95) will be classified as "no
match", coefficients above the second value (default 0.98) will be classified
as "likely match", and coefficients between the two values will be classified
as "possible match".}

\item{quiet}{A logical scalar. Should the function execute quietly, or should
it return status updates throughout the function (default)?}

\item{...}{Additional named arguments to pass to
\code{\link{match_signatures}}.}
}
\value{
The input \code{data} data frame, with an additional \code{match} field which
is a character vector with the following possible values:
\itemize{
\item "match": the original correlation coefficient was greater than or equal to
\code{check_threshold}.
\item "likely match": the original correlation coefficient was less than
\code{check_threshold} and the new coefficient was greater than or equal to
\code{confirm_thresholds[2]}.
\item "possible match: the original correlation coefficient was less than
\code{check_threshold} and the new coefficient was less than
\code{confirm_thresholds[2]} and greater than or equal to
\code{confirm_thresholds[1]}.
\item "no match": the original correlation coefficient was less than
\code{check_threshold} and the new coefficient was less than
\code{confirm_thresholds[1]}.
}
}
\description{
\code{confirm_matches} takes the results of \code{\link{identify_matches}}
and performs an additional analysis on image signature pairs whose
correlations were not high enough to guarantee a match but not low enough to
discount a match. The default implementation assumes that image signatures
were compared using the greyscale components, and so it performs an
additional correlation using the colour components.
}
\details{
By default, \code{confirm_matches} is called with default parameters by
\code{\link{identify_matches}}. It is therefore usually only necessary to
run \code{confirm_matches} on its own in order to change the parameters.
}
\examples{
\dontrun{
# Setup
sigs <- create_signature(test_urls)
matches <- match_signatures(sigs)

# compare_images is only necessary to run if identify_matches was run with `confirm = FALSE`
matches <- identify_matches(matches, confirm = FALSE)
confirm <- confirm_matches(matches, check_threshold = 0.995)
}
}
