% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/download_images.R
\name{download_images}
\alias{download_images}
\title{Download images to a local folder}
\usage{
download_images(
  x = NULL,
  destination,
  path = photos,
  id = id,
  check_duplicates = TRUE,
  quiet = FALSE
)
}
\arguments{
\item{x}{A data frame with a column containing URLs to be downloaded and
(optionally) IDs to be linked to the downloads for file naming and duplicate
detection. If x is supplied, the \code{path} and \code{id} arguments will be
interpreted as field names. If x is NULL, they will be interpreted as the
names of lists or character vectors.}

\item{destination}{A character scalar indicating a local folder where images
should be downloaded to.}

\item{path}{If x is supplied, the name of a list or character field in x with
URLs to be downloaded. If x is NULL, a list or character vector of URLs. In
either case, elements should begin with "http", "https", "ftp" or "ftps". If
the URL has no extension, it will be given the extension ".jpg".}

\item{id}{If x is supplied, the name of a character field in x with IDs to
match to the file downloads. If x is NULL, a character vector the same length
as \code{path}, indicating the IDs of the files to be downloaded. This parameter
can be NULL. If it is supplied, files will be named according to the IDs.}

\item{check_duplicates}{A logical scalar (default TRUE) indicating whether to
skip downloads for files whose IDs match files already present in the
download folder. This argument is ignored if \code{id} is NULL.}

\item{quiet}{A logical scalar. Should the function execute quietly, or should
it return status updates throughout the function (default)?}
}
\value{
Invisibly, a data frame summarizing the results of the downloads.
}
\description{
\code{download_images} takes a data frame or a vector of URLs and downloads
the files to a local folder, possibly with the use of an additional ID vector
to rename the files and check for duplicates.
}
