% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/find_duplicates.R
\name{find_duplicates}
\alias{find_duplicates}
\title{Find and label duplicate image signatures}
\usage{
find_duplicates(x, threshold = 80, find_all = FALSE, quiet = FALSE)
}
\arguments{
\item{x}{A data frame with two columns containing matchr_signature vectors.
If there are columns named \code{x_sig} and \code{y_sig}, these will be the vectors
which will be analyzed. If not, the first two columns containing
matchr_signature vectors will be used. If more than two such columns are
present, a warning will be issued.}

\item{threshold}{A length-one integer vector. Which Hamming distance should
be used to consider images to be identical? If the distance between two \code{x}
or two \code{y} images is <= \code{threshold}, the images will be considered
duplicates.}

\item{find_all}{A logical scalar. Should the function find all \code{y} duplicates
even for rows which do not have \code{x} duplicates (default FALSE)? If FALSE,
rows will be checked for \code{x} duplicates first, and any row without an \code{x}
duplicate will be removed from the search for \code{y} duplicates. This can result
in considerable speed gains if the goal is to find rows which have both \code{x}
and \code{y} duplicates (e.g. for subsequent processing in \code{confirm_matches}).}

\item{quiet}{A logical scalar. Should the function execute quietly, or should
it return status updates throughout the function (default)?}
}
\value{
The input data frame, with additional \code{x_id} and \code{y_id} fields which
identify duplicated image signatures.
}
\description{
\code{find_duplicates} takes a data frame with two matchr_signature vectors
and identifies and labels duplicate signatures, according to a given set of
thresholds.
}
