% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/identify_image.R
\name{identify_image}
\alias{identify_image}
\alias{identify_image.cimg}
\alias{identify_image.character}
\alias{identify_image.list}
\alias{identify_image.logical}
\title{Generate a unique colour signature for an image}
\usage{
identify_image(image, bands = 25, ...)

\method{identify_image}{cimg}(image, bands = 25, ...)

\method{identify_image}{character}(image, bands = 25, mem_limit = 50, quiet = FALSE, ...)

\method{identify_image}{list}(image, bands = 25, quiet = FALSE, ...)

\method{identify_image}{logical}(image, bands = 25, ...)
}
\arguments{
\item{image}{Object (or list of objects) of class 'cimg', probably imported
using \code{load_image} or \code{imager::load.image}, or file path
or URL (or vector/list of file paths or URLs) of an image.}

\item{bands}{Integer scalar. The number of horizontal bands the image should
be split into for processing. Higher values will produce a more distinctive
colour signature, potentially decreasing the rate of matching false
positives, but at the cost of increased processing time and an increased rate
of matching false negatives.}

\item{...}{Additional arguments passed to methods.}

\item{mem_limit}{An integer scalar. How many images should the function load
into memory before extracting image signatures and releasing the associated
memory? Higher values will lead to the function executing more quickly, but
can result in enormous memory requirements. The function uses approximately
10 MB per image, so the default value of 50 will cause peak memory usage of
approximately 5 GB.}

\item{quiet}{A logical scalar. Should the function execute quietly, or should
it return status updates throughout the function (default)?}
}
\value{
A numeric vector of length \code{bands}.
}
\description{
\code{identify_image} takes an image file and generates a numeric vector of
average colour values, so that the image can be compared to others.
}
\details{
The function works as follows. An image is decomposed into horizontal bands
(with the number of bands controlled by the \code{bands} argument), and for each
band an average colour is calculated. The vector of these colour averages
becomes a distinctive signature that can identify a given image even if the
image is rescaled or compressed, and thus serves as a reliable indicator of
whether two images are the same.
}
