% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/identify_matches.R
\name{identify_matches}
\alias{identify_matches}
\alias{identify_matches.matchr_matrix}
\alias{identify_matches.matchr_sig}
\title{Identify matches from a matchr image matrix}
\usage{
identify_matches(
  x,
  y = NULL,
  threshold = 0.975,
  confirm = TRUE,
  quiet = FALSE,
  ...
)

\method{identify_matches}{matchr_matrix}(
  x,
  y = NULL,
  threshold = 0.975,
  confirm = TRUE,
  quiet = FALSE,
  ...
)

\method{identify_matches}{matchr_sig}(
  x,
  y,
  threshold = 0.975,
  confirm = TRUE,
  quiet = FALSE,
  method = "grey",
  compare_ar = TRUE,
  stretch = 1.2,
  mem_scale = 0.2,
  ...
)
}
\arguments{
\item{x}{A vector of class \code{matchr_sig}, containing image signatures produced
by \code{\link{create_signature}}, or \code{matchr_matrix}, containing image
correlation matrices produced by \code{\link{match_signatures}}.}

\item{y}{A vector of class \code{matchr_sig}, containing image signatures produced
by \code{\link{create_signature}}. If x is not a \code{matchr_sig} vector, this
argument is ignored, and will trigger a warning if a non-NULL value is
supplied.}

\item{threshold}{A numeric scalar. The minimum correlation constant to
consider images to be matched.}

\item{confirm}{A logical scalar. Should the function run
\code{\link{confirm_matches}} on the results before returning them (default)?
If TRUE, \code{\link{confirm_matches}} will be called with default
parameters, i.e. under the assumption that the original signature matching
was done using greyscale signatures, and the confirmation should be carried
out with colour signatures with the default tolerance thresholds.}

\item{quiet}{A logical scalar. Should the function execute quietly, or should
it return status updates throughout the function (default)?}

\item{...}{Additional arguments passed to methods.}

\item{method, compare_ar, stretch, mem_scale}{Arguments passed to
\code{\link{match_signatures}}.}
}
\value{
A tibble if {dplyr} is installed or a data frame if not, with one
row per identified match, and the following columns:
\itemize{
\item \code{matrix}: The match's index position in the input \code{matchr_matrix} vector.
\item \code{list_index}, \code{x_index} and \code{y_index}: The match's list, row and column
index positions in the \code{matchr_matrix} vector element (which is represented
as a matrix but is actually a list of matrices).
\item \code{x_file} and \code{y_file}: The file paths for the images which were matched.
\item \code{correlation}: The Pearson correlation coefficient of the two files'
image signatures.
\item \code{match} (if \code{confirm = TRUE}): A character vector indicating match status.
(See \code{\link{confirm_matches}} for details.)
}
}
\description{
\code{identify_matches} analyzes a \code{matchr_matrix} vector of image signature
correlations to identify possible matches. By default it sets a low
threshold for identifying matches, with the assumption that the results will
subsequently be refined using \code{\link{confirm_matches}}.
}
