% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/identify_matches.R
\name{identify_matches}
\alias{identify_matches}
\title{Identify matches from a matchr image matrix}
\usage{
identify_matches(image_matrix, threshold = 0.99)
}
\arguments{
\item{image_matrix}{A matrix of image correlations produced by
\code{imager::match_images}.}

\item{threshold}{A numeric scalar. The minimum correlation constant to
consider to images to be matched.}
}
\value{
TKTK
}
\description{
\code{identify_matches} TKTK
}
\details{
TKTK
}
