% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/integrate_changes.R
\name{integrate_changes}
\alias{integrate_changes}
\title{Integrate changes in image matching}
\usage{
integrate_changes(result, change_table)
}
\arguments{
\item{result}{TKTK}

\item{change_table}{TKTK}
}
\value{
TKTK
}
\description{
\code{integrate_changes} TKTK
}
\details{
TKTK
}
