% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load_image.R
\name{load_image}
\alias{load_image}
\title{Load image from file or URL}
\usage{
load_image(file, quiet = FALSE)
}
\arguments{
\item{file}{A vector of file paths or URLs to be passed to
\code{imager::load.image}.}

\item{quiet}{A logical scalar. Should the function execute quietly, or should
it return status updates throughout the function (default)?}
}
\value{
If the input is a single file path or URL, the output will be a
single object of class 'cimg'. If the input is a vector with length > 1,
the output will be a list of 'cimg' objects with the same length.
}
\description{
\code{load_image} is a vectorized version of \code{imager::load.image}. It
optionally supports parallel processing (via \code{future} and
\code{future.apply}) and progress reporting (via \code{progressr}).
}
\details{
The function is a wrapper around \code{imager::load.image} with three
enhancements. It can take a vector of input paths instead a single path, it
supports parallel processing, and it support progress reporting.
}
