% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/match_images.R
\name{match_images}
\alias{match_images}
\title{Match images based on colour signatures}
\usage{
match_images(x, y = NULL, bands = 25, quiet = FALSE)
}
\arguments{
\item{x, y}{Lists of 'cimg' objects to be matched, or file paths and URLs
pointing to images which can be imported as 'cimg' objects. If \code{y} is missing
(default), each object in \code{x} will be matched against each other object in
\code{x.} If \code{y} is present, each object in \code{x} will be matched against each
object in \code{y}.}

\item{bands}{Integer scalar to be passed to \code{\link{create_signature}}. The
number of horizontal bands the image should be split into for processing.
Higher values will produce a more distinctive colour signature, potentially
decreasing the rate of matching false positives, but at the cost of increased
processing time and an increased rate of matching false negatives.}

\item{quiet}{A logical scalar. Should the function execute quietly, or should
it return status updates throughout the function (default)?}
}
\value{
A correlation matrix. If \code{x} and \code{y} are both present, the matrix
will have \code{length(x)} rows and \code{length(y)} columns, and for the matrix \code{Q}
the cell \code{Q[i, j]} will be the Pearson correlation coefficient between images
\code{x[[i]]} and \code{y[[j]]}. If \code{y} is not present, the matrix will be square, and
the cell \code{Q[i, j]} will be the correlation between images \code{x[[i]]} and
\code{x[[j]]}.
}
\description{
\code{match_images} takes a list of images and produces a correlation matrix
to identify matches.
}
\details{
A function for identifying matching images. The function takes a list of
images (objects of class 'cimg) and compares their colour signatures to find
matches.

The comparison is done by creating colour signatures for each input image
using \code{\link{create_signature}} and then computing the Pearson correlation
coefficient between these signatures. In general, pairs of images which were
identical prior to arbitrary resampling and compression will have correlation
coefficients of at least 0.99.
}
