% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/match_signatures.R
\name{match_signatures}
\alias{match_signatures}
\title{Match images based on colour signatures}
\usage{
match_signatures(
  x,
  y = NULL,
  method = "grey",
  compare_ar = TRUE,
  stretch = 1.2,
  quiet = FALSE
)
}
\arguments{
\item{x, y}{Vectors of class \code{matchr_signature} to be matched. If \code{y} is
missing (default), each object in \code{x} will be matched against each other
object in \code{x.} If \code{y} is present, each object in \code{x} will be matched against
each object in \code{y}.}

\item{method}{A character scalar. Should images be compared by their
greyscale signatures ("grey", "greyscale, "gray", or "grayscale", the
default), their colour signatures ("colour", "color", "rgb", or "RGB), or
both ("both")?}

\item{compare_ar}{A logical scalar. Should signatures only be compared for
images with similar aspect ratios (default)? If TRUE, k-means clustering is
used to identify breakpoints between aspect ratios that maximize
between-group distance and minimize the total number of calculations that the
function needs to execute. (Values of k between 3 and 8 are evaluated.) Image
signatures from \code{x} are split into lists between these break points. This
argument is forced to FALSE if either \code{x} or \code{y} has fewer than 10 non-NA
elements.}

\item{stretch}{A numeric scalar. When \code{compare_ar} is TRUE, in order to catch
matches that would fall across a break point, image signatures from \code{y} are
split into lists between the break point / \code{stretch} (default 1.2) on the
lower bound and the break point * \code{stretch} on the upper bound. Increasing
this value will possibly catch matches between extremely distorted images,
but at the cost of potentially larger numbers of false positives, and
substantially increased processing time.}

\item{quiet}{A logical scalar. Should the function execute quietly, or should
it return status updates throughout the function (default)?}
}
\value{
A vector of class \code{matchr_matrix}, each element of which is a
correlation matrix for the \code{x} and \code{y} signatures falling in a given aspect
ratio range. If \code{x} and \code{y} are both present, each matrix will have
\code{length(x)} rows and \code{length(y)} columns, and for the matrix \code{Q} the cell
\code{Q[i, j]} will be the Pearson correlation coefficient between images \code{x[[i]]}
and \code{y[[j]]}. If \code{y} is not present, each matrix will be square, and the cell
\code{Q[i, j]} will be the correlation between images \code{x[[i]]} and \code{x[[j]]}.
}
\description{
\code{match_signatures} takes one or two lists of images and produces a
correlation matrix to identify matches.
}
\details{
A function for identifying matching images. The function takes one or two
vectors of images signatures (class \code{matchr_signature}) and compares their
colour signatures to find matches.

The comparison is done by creating colour signatures for each input image
using \code{\link{create_signature}} and then computing the Pearson
correlation coefficient between these signatures. In general, pairs of images
which were identical prior to arbitrary resampling and compression will have
correlation coefficients of at least 0.99.

The function can optionally filter images by aspect ratio, so only images
with very similar aspect ratios will be compared. This can remove potential
false positives and possibly speed up function execution, if images are
relatively evenly split between aspect ratios.
}
