% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/matchr-package.R
\docType{package}
\name{matchr-package}
\alias{matchr}
\alias{matchr-package}
\title{matchr: Fast and Reliable Matching of Images}
\description{
Fast and reliable matching of images. It creates unique colour
    signatures for images by finding average colour values for horizontal and
    vertical bands. This allows two or more images to be compared to each other 
    using correlations of colour values, and the comparisons are insensitive to
    file format, resolution, or compression.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/UPGo-McGill/matchr}
  \item Report bugs at \url{https://github.com/UPGo-McGill/matchr/issues}
}

}
\author{
\strong{Maintainer}: David Wachsmuth \email{david.wachsmuth@mcgill.ca} (\href{http://orcid.org/0000-0001-5689-9527}{ORCID})

}
\keyword{internal}
