% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/matchr.R
\name{matchr}
\alias{matchr}
\title{Sequential/parallel matching}
\usage{
matchr(
  dat_case,
  dat_ctrl,
  match_exact = NULL,
  match_inexact = NULL,
  n_controls = 1L,
  split_n = 1L,
  sequential = TRUE,
  seed = NULL
)
}
\arguments{
\item{dat_case}{data.table}

\item{dat_ctrl}{data.table}

\item{match_exact}{character vector}

\item{match_inexact}{character string}

\item{n_controls}{integer}

\item{split_n}{integer}

\item{sequential}{logical}

\item{seed}{integer}
}
\value{
data.table
}
\description{
Sequential/parallel matching
}
\examples{
abs
}
