% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/matchr_options.R
\name{matchr_options}
\alias{matchr_options}
\alias{matchr.options}
\alias{matchr.force_parallel}
\title{Options used in matchr}
\description{
Below are the R options and environment variables that are used by the matchr
package.
}
\section{matchr options}{


\describe{
\item{\option{matchr.force_parallel}:}{A logical scalar. Default future
strategy plan used unless otherwise specified via . This will also
be the future plan set when calling \code{plan("default")}.  If not specified,
this option may be set when the \pkg{future} package is \emph{loaded} if
command-line option \code{--parallel=ncores} (short \verb{-p ncores}) is specified; if
\code{ncores > 1}, then option future.plan is set to \code{multisession}
otherwise \code{sequential} (in addition to option \option{mc.cores} being set
to \code{ncores}, if \code{ncores >= 1}).  If system environment variable
\env{R_FUTURE_PLAN} is set, then that overrides the future plan set by the
command-line option. (Default: \code{sequential})}
}
}

