% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/class_matchr_sig.R
\name{new_matchr_sig}
\alias{new_matchr_sig}
\title{Create a new matchr_sig object}
\usage{
new_matchr_sig(x, file, aspect_ratio)
}
\arguments{
\item{x}{A numeric vector.}

\item{file}{A character string, corresponding to the path or URL of the file
from which the signature has been generated.}

\item{aspect_ratio}{A numeric scalar, giving the aspect ratio of the image.}
}
\value{
An object of class \code{matchr_sig}.
}
\description{
Create a new matchr_sig object
}
