% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/remove_black_bars.R
\name{remove_black_bars}
\alias{remove_black_bars}
\title{Remove horizontal black bars from an image}
\usage{
remove_black_bars(x)
}
\arguments{
\item{x}{Pixel array or length-1 vector of class \code{matchr_image}.}
}
\value{
An array if the input was an array, and a length-1 vector of class
\code{matchr_image} if the input was a \code{matchr_image} vector.
}
\description{
\code{remove_black_bars} takes a single \code{matchr_image} or raw pixel array and
removes black bars at the top and bottom of the image, if any are identified.
}
\examples{
\dontrun{
# Setup
img <- load_image(test_urls[13])

remove_black_bars(img)
}
}
